Split files by class and move into render subdir
