+++ napari/_qt/experimental/render/test_image.py
